fix ch05

+++ go-jvm/ch05/instructions/loads/iload.go
@@ -8,7 +8,7 @@ import (
 type ILOAD struct{ base.Index8Instruction }
 
 func (self *ILOAD) Execute(frame *rtda.Frame) {
-	_iload(frame, uint(self.Index))
+	_iload(frame, self.Index)
 }
 
 type ILOAD_0 struct{ base.NoOperandsInstruction }
